+++ app/src/main/java/io/sikorka/android/ui/accounts/account_export/AccountExportContract.kt
@@ -4,11 +4,17 @@ import io.sikorka.android.mvp.BaseView
 import io.sikorka.android.mvp.Presenter
 
 interface AccountExportView : BaseView {
-  fun showError(code: Long)
+  fun showError(code: Int)
   fun exportComplete()
 
 }
 
 interface AccountExportPresenter : Presenter<AccountExportView> {
-  fun export(accountHex: String, passphrase: String, encryptionPass: String, confirmation: String, path: String)
+  fun export(
+    accountHex: String,
+    passphrase: String,
+    encryptionPass: String,
+    confirmation: String,
+    path: String
+  )
 }
